Almoste completed
Note that the following problems were solved by this commit:
1) tack canary watchpoint trigger
2) No color change for price

With regards to the 1), I was not sure, but I extract out candlestick[]
from a function to a static field.
This affects parameters of obtainSticks() function.
With regards to 2), prevPrice variable was initialized everytime when
ShowCurrentPrice() was called.
This checkin also includes clock time print.
Still a problem remains, "a comma" is not shown in
price tag.