Make this program work again
It seems that the size of Json increased for obtaining price.
It seems the change was made on May 21, 2021.
The change caused the program not working.
The cause was the size of StaticJsonDocument<256>.
So, increased the size to 512.
Other changes were made for safely compile
for other platform than M5Core2.
"if (lowest == highest) {" line around line 726 was made
for eliminate too many serial outputs when error occurs.